Implemented a new type of username masquerading, which is also activated by setting mmailid to non-zero in mts.conf.  The new type is based on the From: header in the message draft rather than the GECOS field of the passwd file like the old type.
With mmailid turned off, when a user specifies a custom From: header in a draft,
it'll be used, but the user's real address will be revealed in the SMTP envelope
From: and in the Sender: header.  With mmailid on, this revelation is disabled.

Also copied to here my new comment from mh.h explaining (if I understand it
correctly) the functionality of struct swit's minchars field.  Also, changed the
minchars for "help" from 4 to 0.  It doesn't make sense to use a positive
integer that's the exact length of the option.  That'll make it print out like
"-(help)".  The parentheses are supposed to go around the minimum part of the
option you need to specify when abbreviating it.  Perhaps someone thought it
would be cute to put "help" in parentheses to imply that you know about it
already as you just specified it be seeing the current output, but I think this
"overloading" of the meaning of the parentheses is awkward.  I need to go around
and change "help" to 0 everywhere, but it's getting too late to do it tonight.

Also added some other explanatory comments to existing code.