Someone (unfortunately original version history has been lost so I don't know who) changed "danh" to "your-id".  That makes my introductory comment "we'll use version 1.0.4 and my mhost.com account, danh, as examples here" make no sense, though.
If danh is going to be "parameterized", then 1.0.4 should be too.  Doesn't make
sense to parameterize one of the changing items and not the other one.  If it's
going to be done, danh should be changed to YOUR-ID or something and 1.0.4 (and
1_0_4) should be changed to X.Y.Z (and X_Y_Z).  I just thought it'd be clearer
to use real values and disclaim that fact up front.  Anyhow, putting danh back
until such time as someone wants to parameterize both items (and change the
introductory text appropriately).