Fix used-uninitialised warning in new.c
Fix a used-uninitialised warning in new.c by having the check_folders()
routine return first=last=NULL for empty folder lists. This is a more
sensible API anyway. (It would not actually have been possible for
the code to try to use last while it was uninitialised, as it happens.)

Signed-off-by: Peter Maydell <pmaydell@chiark.greenend.org.uk>