Rationalize .TH manpage macro invocations.
This patch addresses the following issues with .TH macro usage:

  .TH COMP %manext1% "%nmhdate%" MH.6.8 [%nmhverion%]

1) "MH.6.8" is no longer relevant.

2) %nmhdate% is a hardwired value that only gets updated along with
   %nmhversion% and is therefore redundant.

This patch nukes the MH.6.8 tag, and replaces the hardwired
%nmhdate% field with the latest date the associated manpage has
been updated.  I pulled the dates - as best I could - from the
git logs.

Going forward, it is incumbent on developers to update the
manpage date stamps as they update the pages themselves.