Garbage collect "normalize".
The code to normalize addresses (look up the hostnames in DNS for the
"official" hostnames) was disabled when we made DUMB the default configuration
in 1.5.  Since them, the -normalize switch was still around in various
utilities, but it didn't do anything.  Remove all of those switches and
garbage collect various other bits of the "normalize" functionality.

At the same time, improve getm() a bit; it now takes an argument specifying
the length of the error buffer.