mhshow: fix display of alternatives after certain nested multiparts
adjust return codes so that we now have a clear indication of whether
there was a hard error (NOTOK), a failure to display due to type
mismatch (DONE), or complete success (OK).  use those codes in
show_multi_internal() to decide whether we're finished with this
multipart or not.

prior to this change, it was impossible to show part 1.2 of a
message constructed like the following, because the disinction
between (the new, as described above) DONE and OK states was
lost while processing multipart/related part 1.1.
 msg part  type/subtype              size description
  28       multipart/mixed           1040
     1     multipart/alternative      623
     1.1   multipart/related          275
     1.1.1 text/html                  105
     1.2   text/plain                  33
     2     text/plain                  83

this fix is related to 9cc2510 and c9794733.  hopefully it's right
this time.