Add failing test case for single-character header parsing. `scan -format '%{x}'' fails to find single-character header `x' in some cases, depending on the position of the header in the file.  Add a simple test case, scan/test-header-parsing, for this that fails.  Alter the test harness to allow a test to indicate it expects to fail.  When it does, that's ignored.  If it passes, that's then an error instead.
Did this because I don't know if or when it will be fixed and didn't
want it to be forgotten in the meantime.  When it is fixed, the test can
be enhanced to cover many more possibilities.  First noticed it with ad
hoc testing of my fewer-lseek change, but it is present in the released
1.6.