Alter netrc permissions error message to match code.  Add test.
The code wants `go=', but the error message, and FAQ, said to clear read
permission from group and other.  Add test to inc/test-pop as that is
the existing netrc test location, but it might be better in its own test
file as other tests for netrc parsing could be added.