Fix test-header-parsing test script.
It was using sh's backticks that chomped the last linefeed!
Add comment referencing the commit that fixed the code this is testing.
Have fewer test cases, sticking to just one or two character headers
with empty or one-character bodies.  Avoids slowing the tests too much
and the fixed bug was to do with hashing one-character headers.