Add trunccpy(), and the convenience TRUNCCPY().
truncpy copies at most size - 1 chars from non-NULL src to non-NULL
dst, and ensures dst is NUL terminated.  If size is zero then it
aborts as dst cannot be NUL terminated.

It's to be used when truncation is intended and correct, e.g.
reporting a possibly very long external string back to the user.  One
of its advantages over strncpy(3) is it doesn't pad in the common
case of no truncation.