test/common.sh.in: Add test_exit() for $? checking.
The stdout and stderr of the command being tested are saved but ignored,
only regurgitated if the test fails.  The test is stated as a fragment
of a test(1) condition, e.g. '-eq 1'.  This is mainly intended for the
quick one-liners to increase gcov coverage and thus valgrind's scope.
Over time, better tests that check more of the command's behaviour can
replace these, keeping the coverage level.