Changes to nmh.man
This adds a reference to the 'COMMANDS' section, corrects - and
hopefully simplifies - a couple of things, and tries to untangle
the line beginning 'show displays...' in the seventh paragraph.

The line 'The first, previous, next or last messages, if they exist.'
doesn't seem to accurately cover the accompanying 'foo:N' listing
in the following, or am I missing something?

first:N
prev:N
next:N
last:N    The first, previous, next or last messages, if they exist.