Changes to comp.man
Remove the slightly legalistic 'in truth of fact', and bring the
mention of man whatnow into the preceding sentence.

The line 'If the draft already exists, comp will ask you as to the
disposition of the draft.' makes no sense to me, and I'm struggling
to make it make sense. If we put to one side the fact that I think
it's a broken sentence, I still haven't managed to get comp to
accept - or prompt me for - a 'replace' option when 'the draft
already exists', so I'm afraid I have to admit defeat here and seek
the wisdom of the elders.