Changes to mh-format.man
This is a beast, and we may not tame it at the first attempt.

Mostly language simplification, some grammar and formatting, trailing
whitespace, that sort of thing. There are one or two instances where
I've taken (small) liberties with language, in favour of 'comprehension
at first glance'.

There are some commented lines hanging around in this. I've left
them in, but I think they should go at the first opportunity.