A cleanup, and possibly some controversy: I've replaced
	Rather than standard message skeleton

with

	User message skeleton

My thinking here is that a) we can do away with the kinda weird
'Rather than', b) I find the distinction between 'default' and
'standard' to be vague and c) it only appears in a few pages
and is easily remedied. It turns up in comp(1) as "An alternative
to the standard skeleton.", where 'standard' is used in the same
way as 'default' is used in rcvdist(1), so I think it's worth
settling on something 'better' and applying it consistently.

My proposal is 'Default foo', for the foo in %nmhetcdir% and
'User foo' for the foo in the user's <mh-dir>.