Changes to folder.man
Kill a couple of empty .PPs, resurrect a lost .SS, a little bit
of slash and burn, stray whitespace, the usual.

As far as I know, which is not far, the \- construct is only
required for \-args, not, e.g., sub\-folder.

I've changed instances of 'subfolder' to sub-folder.

I've changed the 'the' in

	These folders are all preceded by the read-only...

to 'any'.