mhfixmsg: Use folder name to open second message of a sequence.
`mhfixmsg foo' where foo was sequence of more than one message would
sucessfully process the lowest numbered message of the sequence, but
fail to open(2) the second one as the folder was missing from the path.
For example, /home/ralph/mail/inbox/1 was used to open the first, but
/home/ralph/mail/2 was attempted for the second.  I guess this was
because a static array was being trampled during the first's processing,
though didn't try too hard to find where.  Making a copy of it fixed the
problem.