Remove unneeded `\&' from man pages.
The zero-width character is used before a command character that is
intended to be literal but may end up at the start of a line.
Similarly, it's used after an end-of-sentence character when it may end
up at the end of the line but isn't the end of a sentence.
Remove the other cases because they're noise that confuses the issue of
when `\&' should be used.