man/*.man: Use italic for emphasis, not bold or SHOUTING.
When reading a body of text, italic is sufficient to add emphasis to
what's already being read.  Bold draws the eye to the word on the page
and is more useful for keywords, concepts, etc.  Some `.B' remain that
should probably be `.I', but they looked to have other problems so I
left them out of this pass.