man/*.man: Fix some multi-word .B invocations.
Some needed to be .B for the first word, but .I for the second, or just
plain text for the rest of the words that should be on their own line.
Others were just plain wrong and became .IR for a man-page reference.
Then the .TP macro uses an input trap so only a single following line of
source is used and this prevents a two-line `.B .I', so use embedded
font escapes instead.