Remove leading and trailing blank lines from code.
Many of the main source file start with a blank line, and many end with
one or more too.  They seem to serve little purpose so delete them.
This makes all the files consistent in not having them and ideally the
first line of a C source file will say something about its content in
time instead of just "/*".  Avoided editing format files, etc,. where
the line could be significant, and historical files.