tws.h: Remove zero-valued TW_SNIL macro, used once.
Macro TW_SNIL was the zero value for a two-bit field.  It was only used
once, and that is better ordered to check for bits set with none set
being the last, else, case, rather than in the middle.  Kept the
behaviour of undefined value 3 being treated as TW_SIMP, 2.