sbr/vector.c: calloc(3) rather than malloc(3) and memset(3).
calloc(3) tells libc upfront that the memory needs to be zeroed rather
than giving it the news later with memset.  Perhaps this allows it to
allocate from a CoW page of zero bytes, or it benefits from some other
way in memset not having to clear each bit, but it knocks about 15% off
the wall-clock time and the number of library calls under ltrace(1).