config/version.sh: Rewrite.  Use uname(1), git-describe(1), and UTC.
Not sure why it searched through PATH manually for uname(1) and
hostname(1).  uname and its -n option are POSIX so just use those.  It
was the preference over hostname anyway.

Use git-describe(1), not just git-branch(1), as it gives more detail,
including --dirty to show the built source differs from the commit.

Specify the format for the build date, and its timezone; +0000.