sbr/fmt_scan.c: Fix `foo%-42{bar}' right-justification bug.
The space padding was added at the start of the output buffer, not the
start of the component being formatted.  Caused by 92128dac's move to
dynamic allocation for fmt_scan()'s output.  Only shows if the component
isn't at the start of the buffer.  Expand existing
test/format/test-rightjustify to cover this.