Makefile.am: Alter long lists to be sorted, one entry per line.
It was hard to see what files were included in some variables when
searching for the best place for new content.  Some variables had
definitions that were almost sorted, by weren't.  Switch to a
one-entry-per-line list, in `LC_ALL=C sort' order, documented at the
start of the file.  Makes it easy to run through the lists and see the
patterns.