uip/mhstoresbr.c: Fix single-character nmh-storage bug.
If the profile component nmh-storage was a single character, e.g. the
relative directory `d', then it would be dropped from the output path so
instead of `d/foo' being written, `/foo' would probably fail.  I think
this was due to a faulty check for nmh-storage being `/'.  Add a test
for a non-/ single-character nmh-storage.