uip/mhcachesbr.c: Move exported symbols to existing h/mhcachesbr.h.
Remove the user's copies of the declarations.
A static struct swit caches[] was declared everywhere that included
h/mhcachesbr.h.  The larger number of includers causes `variable not
used' for `caches'.  Move that declaration to uip/mhcachesbr.c and have
an exported `cache_policy' symbol that points to it for the existing
users.