Fix bug in cpnumber().
The cpnumber() function (which handles the NUMF instruction, among others)
would hang if a 0 width was given to it.  Make sure that (and negative
widths) are handled correctly.  Note that normally NUMF did not handle
left padding which is indicated by a negative width; that may change
in the future.