Makefile.am: Remove multiple `mkdir -p' for etc;  use ./configure.
A `mkdir -p' was used in each rule that created a file in etc to ensure
the directory already existed.  Ken pointed out existing ./configure
code to do that for the man directory, just once.  Use that for etc too.
Remove the `test -d' because `mkdir -p' does that itself.